% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport.R
\name{generer_rapport}
\alias{generer_rapport}
\title{Générer un rapport personnalisé pour une commune et un département}
\usage{
generer_rapport(commune, departement, output)
}
\arguments{
\item{commune}{Un entier spécifiant le code INSEE ou postal de la commune pour laquelle le rapport sera généré.}

\item{departement}{Un entier spécifiant le code du département pour lequel le rapport sera généré.}

\item{output}{Le chemin de sortie où le fichier généré (par exemple, PDF) sera enregistré. Ce doit être un
chemin de fichier valide.}
}
\value{
Aucun retour explicite. Le rapport est généré et enregistré à l'emplacement spécifié dans \code{output}.
}
\description{
Cette fonction génère un rapport personnalisé au format PDF en utilisant un fichier Quarto (.qmd).
Elle prend en paramètre les informations sur la commune (identifiée par son code) et le département
(identifié par son code) et génère un rapport en utilisant un modèle défini dans le fichier \code{rapport.qmd}
situé dans le dossier \code{inst} du package.
}
\examples{
# Exemple d'utilisation pour générer un rapport pour Nantes (commune 44109) et Loire-Atlantique (département 44)
generer_rapport(commune = 44109, departement = 44, output = "rapport_nantes.pdf")

# Exemple avec un autre département et commune
generer_rapport(commune = 75056, departement = 75, output = "rapport_paris.pdf")

}
