% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.commune.R
\name{plot.commune}
\alias{plot.commune}
\title{Visualiser la repartition des elus par categorie socio-professionnelle dans une commune}
\usage{
\method{plot}{commune}(df, ...)
}
\arguments{
\item{df}{Un data frame contenant des informations sur les elus, incluant les colonnes \verb{Code de la catégorie socio-professionnelle},
\verb{Libellé de la catégorie socio-professionnelle}, \verb{Libellé de la commune} et \verb{Libellé du département}.}

\item{...}{Arguments supplementaires, non utilises dans cette methode.}
}
\value{
Un graphique genere avec \code{ggplot2} representant la repartition des elus de la commune par categorie socio-professionnelle.
Il affiche les codes professionnels en x et le nombre d'elus par categorie sur l'axe y.
}
\description{
Cette fonction genere un graphique montrant la repartition des elus d'une commune par categorie
socio-professionnelle. Le graphique est un diagramme a barres, ou l'axe des x represente les codes
des categories socio-professionnelles et l'axe des y le nombre d'elus dans chaque categorie.
La fonction utilise \code{ggplot2} pour afficher les resultats.
}
\examples{
# Exemples d'utilisation


# Donnees (villes)

df_Nantes <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Nantes")

df_Aignan <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Aignan")


# Utilisation de la fonction

plot.commune(df_Nantes)

plot.commune(df_Aignan)

}
