% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.departement.R
\name{plot.departement}
\alias{plot.departement}
\title{Visualiser la répartition des élus par catégorie socio-professionnelle dans un département}
\usage{
\method{plot}{departement}(df, ...)
}
\arguments{
\item{df}{Un data frame contenant des informations sur les élus, incluant les colonnes \verb{Code de la catégorie socio-professionnelle},
\verb{Libellé de la catégorie socio-professionnelle}, \verb{Libellé du département}, et \verb{Code de la commune}.}

\item{...}{Arguments supplémentaires, non utilisés dans cette méthode.}
}
\value{
Un graphique généré avec \code{ggplot2} représentant la répartition des élus du département par catégorie socio-professionnelle.
Il affiche les 10 catégories socio-professionnelles les plus représentées en fonction du nombre d'élus.
}
\description{
Cette fonction génère un graphique montrant la répartition des élus d'un département par catégorie
socio-professionnelle. Le graphique est un diagramme à barres, où l'axe des x représente les codes
des catégories socio-professionnelles et l'axe des y le nombre d'élus dans chaque catégorie.
Elle affiche les 10 catégories les plus représentées dans le département.
}
\examples{
# Exemples d'utilisation


# Données (départements)

df_Loire_Atlantique <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Loire-Atlantique")

df_Gers <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Gers")


# Data frames appartenant à la classe departement

df_Loire_Atlantique <- structure(df_Loire_Atlantique, class = c("departement", class(df_Loire_Atlantique)))

df_Gers <- structure(df_Gers, class = c("df_Gers", class(df_Gers)))


# Utilisation de la fonction

plot.departement(df_Loire_Atlantique)

plot.departement(df_Gers)

}
