% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_code_professions.R
\name{plot_code_professions}
\alias{plot_code_professions}
\title{Visualiser la répartition des élus par catégorie socio-professionnelle}
\usage{
plot_code_professions(df)
}
\arguments{
\item{df}{Un DataFrame contenant une colonne \verb{Code de la catégorie socio-professionnelle}.}
}
\value{
Un objet ggplot affichant un diagramme en barres de la distribution des élus
par catégorie socio-professionnelle.
}
\description{
Cette fonction génère un graphique en barres représentant la répartition
des élus selon leur catégorie socio-professionnelle.
}
\examples{
# Exemples d'utilisation


# Données (villes ou départements)

df_Nantes <- df_gers_loire_atlantique |>
  filter(`Libellé de la commune` == "Nantes")

df_Aignan <- df_gers_loire_atlantique |>
  filter(`Libellé de la commune` == "Aignan")

df_Loire_Atlantique <- df_gers_loire_atlantique |>
  filter(`Libellé du département` == "Loire-Atlantique")

df_Gers <- df_gers_loire_atlantique |>
  filter(`Libellé du département` == "Gers")


# Utilisation de la fonction

plot_code_professions(df_Nantes)

plot_code_professions(df_Aignan)

plot_code_professions(df_Loire_Atlantique)

plot_code_professions(df_Gers)
}
\keyword{internal}
