% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.commune.R
\name{summary.commune}
\alias{summary.commune}
\title{Resumer les informations d'une commune}
\usage{
\method{summary}{commune}(df, ...)
}
\arguments{
\item{df}{Un data frame contenant des informations sur les elus d'une commune,
incluant les colonnes \verb{Libellé de la commune} et \verb{Date de naissance}.}

\item{...}{Arguments supplementaires, non utilises dans cette methode.}
}
\value{
Aucun retour explicite. La fonction affiche dans la console :
\itemize{
\item Le nom de la commune,
\item Le nombre d’elus,
\item La distribution des ages des elus sous forme de tibble,
\item Le nom, le prenom, la date de naissance et l'age de l’elu(e) le/la plus age(e), sous forme de tibble.
}
}
\description{
Cette fonction affiche un resume des informations d'une commune,
notamment le nom de la commune, le nombre d'elus, la distribution des ages
des elus, ainsi que le nom et l'age de l'elu(e) le/la plus age(e).
}
\examples{
# Exemples d'utilisation


# Donnees (villes)

data("df_Gers_Loire_Atlantique")

df_Nantes <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Nantes")

df_Aignan <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Aignan")


# Utilisation de la fonction

summary.commune(df_Nantes)

summary.commune(df_Aignan)

}
