% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.departement.R
\name{summary.departement}
\alias{summary.departement}
\title{Resumer les informations d'un departement}
\usage{
\method{summary}{departement}(df, ...)
}
\arguments{
\item{df}{Un data frame contenant des informations sur les elus,
incluant les colonnes \verb{Libellé du département}, \verb{Code de la commune}, et les informations necessaires
pour calculer la distribution des ages, trouver l’elu(e) le/la plus age(e) et le/la plus jeune,
ainsi que la moyenne d’age par commune.}

\item{...}{Arguments supplementaires, non utilises dans cette methode.}
}
\value{
Aucun retour explicite. La fonction affiche dans la console :
\itemize{
\item Le nom du departement,
\item Le nombre de communes dans le departement,
\item Le nombre d’elus dans le departement,
\item La distribution des ages des elus sous forme de tibble,
\item Le nom, le prenom, la date de naissance et l'age de l’elu(e) le/la plus age(e),
ainsi que celui/ceux de l’elu(e) le/la plus jeune, sous forme de tibble,
\item Les noms des communes avec la moyenne d’age la plus faible et la plus elevee,
ainsi que la distribution des ages des elus pour ces communes.
}
}
\description{
Cette fonction affiche un resume des informations d'un departement,
notamment le nom du departement, le nombre de communes, le nombre d'elus,
la distribution des ages des elus, ainsi que le nom et l'age de l'elu(e) le/la plus age(e)
et de l'elu(e) le/la plus jeune. Elle affiche egalement les communes ayant la moyenne d’age
la plus faible et la plus elevee, avec la distribution des ages des elus pour ces communes.
}
\examples{
# Exemples d'utilisation


# Donnees (departements)

df_Loire_Atlantique <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Loire-Atlantique")

df_Gers <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Gers")


# Utilisation de la fonction

summary.departement(df_Loire_Atlantique)

summary.departement(df_Gers)

}
