% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver_elu_le_plus_age_jeune.R
\name{trouver_elu_le_plus_age_jeune}
\alias{trouver_elu_le_plus_age_jeune}
\title{Trouver l'elu(e) le plus age(e) et le/la plus jeune}
\usage{
trouver_elu_le_plus_age_jeune(df)
}
\arguments{
\item{df}{Un data frame contenant les informations sur les elus,
incluant les colonnes \verb{Nom de l'élu}, \verb{Prénom de l'élu}, \verb{Date de naissance},
et \verb{Libellé de la commune}.}
}
\value{
Un tibble avec les informations de l'elu(e) le/la plus age(e)
et de l'elu(e) le/la plus jeune. Les colonnes retournees incluent :
\itemize{
\item \verb{Nom de l'élu} : Le nom de l'elu,
\item \verb{Prénom de l'élu} : Le prenom de l'elu,
\item \verb{Date de naissance} : La date de naissance de l'elu,
\item \code{âge} : L'age calculé de l'elu,
\item \verb{Libellé de la commune} : La commune a laquelle appartient l'elu.
}
}
\description{
Cette fonction permet de trouver l'elu(e) le plus age(e) et le/la plus jeune
dans un data frame d'elus. Elle calcule l'age des elus a partir de leur date de naissance,
puis retourne les informations concernant l'elu(e) le/la plus age(e) et le/la plus jeune.
La fonction trie les elus par date de naissance et selectionne le premier et le dernier elu.
}
\examples{
# Exemples d'utilisation


# Donnees (villes ou departements)

data("df_Gers_Loire_Atlantique")

df_Nantes <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Nantes")

df_Aignan <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Aignan")

df_Loire_Atlantique <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Loire-Atlantique")

df_Gers <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Gers")


# Utilisation de la fonction

trouver_elu_le_plus_age_jeune(df_Nantes)

trouver_elu_le_plus_age_jeune(df_Aignan)

trouver_elu_le_plus_age_jeune(df_Loire_Atlantique)

trouver_elu_le_plus_age_jeune(df_Gers)
}
