% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver_moyenne_age_faible_elevee.R
\name{trouver_moyenne_age_faible_elevee}
\alias{trouver_moyenne_age_faible_elevee}
\title{Trouver les communes avec la moyenne d'age la plus faible et la plus elevee}
\usage{
trouver_moyenne_age_faible_elevee(df)
}
\arguments{
\item{df}{Un data frame contenant les informations sur les elus,
incluant les colonnes \verb{Libellé de la commune} et les informations necessaires
pour calculer la moyenne d'age des elus par commune.}
}
\value{
Un tibble contenant les informations des communes avec la moyenne d'age
la plus faible et la plus elevee. Les colonnes retournees incluent :
\itemize{
\item \verb{Libellé de la commune} : Le nom de la commune,
\item \code{Moyenne} : La moyenne d'age des elus dans la commune.
}
}
\description{
Cette fonction permet de trouver les communes avec la moyenne d’age des elus la plus faible
et la plus elevee. Elle regroupe les donnees par commune, applique une fonction de calcul
de la distribution des ages des elus à chaque groupe, puis selectionne les communes avec
les moyennes d'age les plus basses et les plus elevees.
}
\examples{
# Exemples d'utilisation


# Donnees (villes ou departements)

data("df_Gers_Loire_Atlantique")

df_Nantes <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Nantes")

df_Aignan <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Aignan")

df_Loire_Atlantique <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Loire-Atlantique")

df_Gers <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Gers")


# Utilisation de la fonction

trouver_moyenne_age_faible_elevee(df_Nantes)

trouver_moyenne_age_faible_elevee(df_Aignan)

trouver_moyenne_age_faible_elevee(df_Loire_Atlantique)

trouver_moyenne_age_faible_elevee(df_Gers)
}
