% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_distribution_age.R
\name{calcul_distribution_age}
\alias{calcul_distribution_age}
\title{Calculer la distribution des âges}
\usage{
calcul_distribution_age(df, ...)
}
\arguments{
\item{df}{Un data frame contenant une colonne \verb{Date de naissance} et
une colonne \verb{Libellé de la commune} ou \verb{Libellé du département}.}

\item{...}{Paramètres supplémentaires (non utilisés directement dans cette fonction).}
}
\value{
Un tibble contenant le nom de la commune ou du département, ainsi que
les statistiques d'âge : minimum, premier quartile (Q1), médiane, moyenne,
troisième quartile (Q3) et maximum.
}
\description{
Cette fonction calcule la distribution des âges des individus d'un data frame
en renvoyant des statistiques descriptives : minimum, quartiles, médiane, moyenne et maximum.
}
\examples{
# Exemples d'utilisation


# Données (villes ou départements)

data("df_Gers_Loire_Atlantique")

df_Nantes <- df_Gers_Loire_Atlantique |>
  dplyr::filter(`Libellé de la commune` == "Nantes")

df_Aignan <- df_Gers_Loire_Atlantique |>
  dplyr::filter(`Libellé de la commune` == "Aignan")

df_Loire_Atlantique <- df_Gers_Loire_Atlantique |>
  dplyr::filter(`Libellé du département` == "Loire-Atlantique")

df_Gers <- df_Gers_Loire_Atlantique |>
  dplyr::filter(`Libellé du département` == "Gers")


# Utilisation de la fonction

calcul_distribution_age(df_Nantes)

calcul_distribution_age(df_Aignan)

calcul_distribution_age(df_Loire_Atlantique)

calcul_distribution_age(df_Gers)
}
