% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compter_nombre_d_adjoints.R
\name{compter_nombre_d_adjoints}
\alias{compter_nombre_d_adjoints}
\title{Compter le nombre d'adjoints}
\usage{
compter_nombre_d_adjoints(df)
}
\arguments{
\item{df}{Un DataFrame contenant une colonne \verb{Libellé de la fonction} (type \code{character}).}
}
\value{
Un entier représentant le nombre d'occurrences du mot "adjoint".
}
\description{
Cette fonction compte le nombre d'occurrences du mot "adjoint"
dans la colonne \verb{Libellé de la fonction} d'un DataFrame.
}
\examples{
# Exemples d'utilisation


# Données (villes ou départements)

df_Nantes <- df_gers_loire_atlantique |>
  filter(`Libellé de la commune` == "Nantes")

df_Aignan <- df_gers_loire_atlantique |>
  filter(`Libellé de la commune` == "Aignan")

df_Loire_Atlantique <- df_gers_loire_atlantique |>
  filter(`Libellé du département` == "Loire-Atlantique")

df_Gers <- df_gers_loire_atlantique |>
  filter(`Libellé du département` == "Gers")


# Utilisation de la fonction

compter_nombre_d_adjoints(df_Nantes)

compter_nombre_d_adjoints(df_Aignan)

compter_nombre_d_adjoints(df_Loire_Atlantique)

compter_nombre_d_adjoints(df_Gers)
}
\keyword{internal}
