% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compter_nombre_d_elus.R
\name{compter_nombre_d_elus}
\alias{compter_nombre_d_elus}
\title{Compter le nombre d'élus}
\usage{
compter_nombre_d_elus(df)
}
\arguments{
\item{df}{Un DataFrame contenant les colonnes \verb{Nom de l'élu}, \verb{Prénom de l'élu}, et \verb{Date de naissance} (type \code{character} pour les deux premières et \code{Date} pour la dernière).}
}
\value{
Un entier représentant le nombre d'élus distincts dans le DataFrame.
}
\description{
Cette fonction compte le nombre d'élus distincts dans un DataFrame en se basant
sur les colonnes \verb{Nom de l'élu}, \verb{Prénom de l'élu}, et \verb{Date de naissance}.
}
\examples{
# Exemples d'utilisation


# Données (villes ou départements)

df_Nantes <- df_gers_loire_atlantique |>
  filter(`Libellé de la commune` == "Nantes")

df_Aignan <- df_gers_loire_atlantique |>
  filter(`Libellé de la commune` == "Aignan")

df_Loire_Atlantique <- df_gers_loire_atlantique |>
  filter(`Libellé du département` == "Loire-Atlantique")

df_Gers <- df_gers_loire_atlantique |>
  filter(`Libellé du département` == "Gers")


# Utilisation de la fonction

compter_nombre_d_elus(df_Nantes)

compter_nombre_d_elus(df_Aignan)

compter_nombre_d_elus(df_Loire_Atlantique)

compter_nombre_d_elus(df_Gers)
}
\keyword{internal}
