% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_commune.R
\name{creer_commune}
\alias{creer_commune}
\title{Creer un objet de classe 'commune' a partir d'un data frame}
\usage{
creer_commune(df)
}
\arguments{
\item{df}{Un data frame contenant des informations sur les elus d'une commune. Le data frame doit inclure la colonne
\verb{Code de la commune}.}
}
\value{
Un objet de classe \code{commune} contenant les memes informations que le data frame d'entree, mais avec une nouvelle
classe attribuee.
}
\description{
Cette fonction prend un data frame contenant des informations sur une commune et valide que toutes les lignes correspondent
a une seule commune (en verifiant que le \verb{Code de la commune} est unique). Elle transforme ensuite ce data frame en un objet de
classe \code{commune} pour lui attribuer un comportement specifique lors de son utilisation.
}
\examples{
# Exemples d'utilisation


# Donnees (villes)

data("df_Gers_Loire_Atlantique")

df_Nantes <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Nantes")

df_Aignan <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Aignan")


# Utilisation de la fonction

df_Nantes <- creer_commune(df_Nantes)

df_Aignan <- creer_commune(df_Aignan)

}
