% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_departement.R
\name{creer_departement}
\alias{creer_departement}
\title{Créer un objet de classe 'departement' à partir d'un data frame}
\usage{
creer_departement(df)
}
\arguments{
\item{df}{Un data frame contenant des informations sur les élus d'un département. Le data frame doit inclure la colonne
\verb{Code du département}.}
}
\value{
Un objet de classe \code{departement} contenant les mêmes informations que le data frame d'entrée, mais avec une nouvelle
classe attribuée.
}
\description{
Cette fonction prend un data frame contenant des informations sur un département et valide que toutes les lignes correspondent
à un seul département (en vérifiant que le \verb{Code du département} est unique). Elle transforme ensuite ce data frame en un objet de
classe \code{departement} pour lui attribuer un comportement spécifique lors de son utilisation.
}
\examples{
# Exemples d'utilisation


# Données (départements)

data("df_Gers_Loire_Atlantique")

df_Loire_Atlantique <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Loire-Atlantique")

df_Gers <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Gers")


# Utilisation de la fonction

df_Loire_Atlantique <- creer_departement(df_Loire_Atlantique)

df_Gers <- creer_departement(df_Gers)

}
