% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport.R
\name{generer_rapport}
\alias{generer_rapport}
\title{Générer un rapport Quarto et le déplacer vers un chemin spécifié}
\usage{
generer_rapport(commune, departement, output)
}
\arguments{
\item{commune}{Un entier représentant le numéro de la commune à inclure dans le rapport.}

\item{departement}{Un entier représentant le numéro du département à inclure dans le rapport.}

\item{output}{Une chaîne de caractères représentant le chemin de destination où le fichier rapport doit être déplacé. Le fichier généré sera déplacé vers ce chemin sous le nom de \code{rapport.html}.}
}
\value{
Aucun retour explicite. Le fichier \code{rapport.html} est généré et déplacé vers le chemin spécifié.
}
\description{
Cette fonction génère un rapport Quarto en utilisant les paramètres \code{commune} et \code{departement} spécifiés.
Après la génération, le fichier \code{rapport.html} est déplacé vers le chemin de destination spécifié par l'utilisateur.
}
\examples{
# Utilisation de la fonction pour générer un rapport pour la commune de Nantes (44109) et le département de Loire-Atlantique (44)

generer_rapport(44109, 44, "<votre chemin>/rapport_final.html")

}
