% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.commune.R
\name{plot.commune}
\alias{plot.commune}
\title{Visualiser la répartition des élus par catégorie socio-professionnelle dans une commune}
\usage{
\method{plot}{commune}(df, ...)
}
\arguments{
\item{df}{Un data frame contenant des informations sur les élus, incluant les colonnes \verb{Code de la catégorie socio-professionnelle},
\verb{Libellé de la catégorie socio-professionnelle}, \verb{Libellé de la commune} et \verb{Libellé du département}.}

\item{...}{Arguments supplémentaires, non utilisés dans cette méthode.}
}
\value{
Un graphique généré avec \code{ggplot2} représentant la répartition des élus de la commune par catégorie socio-professionnelle.
Il affiche les codes professionnels en x et le nombre d'élus par catégorie sur l'axe y.
}
\description{
Cette fonction génère un graphique montrant la répartition des élus d'une commune par catégorie
socio-professionnelle. Le graphique est un diagramme à barres, où l'axe des x représente les codes
des catégories socio-professionnelles et l'axe des y le nombre d'élus dans chaque catégorie.
La fonction utilise \code{ggplot2} pour afficher les résultats.
}
\examples{
# Exemples d'utilisation


# Données (villes)

data("df_Gers_Loire_Atlantique")

df_Nantes <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Nantes")

df_Aignan <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Aignan")


# Utilisation de la fonction

plot.commune(df_Nantes)

plot.commune(df_Aignan)

}
