% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.departement.R
\name{plot.departement}
\alias{plot.departement}
\title{Visualiser la repartition des elus par categorie socio-professionnelle dans un departement}
\usage{
\method{plot}{departement}(df, ...)
}
\arguments{
\item{df}{Un data frame contenant des informations sur les elus, incluant les colonnes \verb{Code de la catégorie socio-professionnelle},
\verb{Libellé de la catégorie socio-professionnelle}, \verb{Libellé du département}, et \verb{Code de la commune}.}

\item{...}{Arguments supplementaires, non utilises dans cette methode.}
}
\value{
Un graphique genere avec \code{ggplot2} representant la repartition des elus du departement par categorie socio-professionnelle.
Il affiche les 10 categories socio-professionnelles les plus representees en fonction du nombre d'elus.
}
\description{
Cette fonction genere un graphique montrant la repartition des elus d'un departement par categorie
socio-professionnelle. Le graphique est un diagramme a barres, ou l'axe des x represente les codes
des categories socio-professionnelles et l'axe des y le nombre d'elus dans chaque categorie.
Elle affiche les 10 categories les plus representees dans le departement.
}
\examples{
# Exemples d'utilisation


# Donnees (departements)

df_Loire_Atlantique <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Loire-Atlantique")

df_Gers <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Gers")


# Utilisation de la fonction

plot.departement(df_Loire_Atlantique)

plot.departement(df_Gers)

}
