% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.commune.R
\name{summary.commune}
\alias{summary.commune}
\title{Résumer les informations d'une commune}
\usage{
\method{summary}{commune}(df, ...)
}
\arguments{
\item{df}{Un data frame contenant des informations sur les élus d'une commune,
incluant les colonnes \verb{Libellé de la commune} et \verb{Date de naissance}.}

\item{...}{Arguments supplémentaires, non utilisés dans cette méthode.}
}
\value{
Aucun retour explicite. La fonction affiche dans la console :
\itemize{
\item Le nom de la commune,
\item Le nombre d’élus,
\item La distribution des âges des élus sous forme de tibble,
\item Le nom, le prénom, la date de naissance et l'âge de l’élu(e) le/la plus âgé(e), sous forme de tibble.
}
}
\description{
Cette fonction affiche un résumé des informations d'une commune,
notamment le nom de la commune, le nombre d'élus, la distribution des âges
des élus, ainsi que le nom et l'âge de l'élu(e) le/la plus âgé(e).
}
\examples{
# Exemples d'utilisation


# Données (villes)

df_Nantes <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Nantes")

df_Aignan <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Aignan")


# Data frames appartenant à la classe commune

df_Nantes <- structure(df_Nantes, class = c("commune", class(df_Nantes)))

df_Aignan <- structure(df_Aignan, class = c("commune", class(df_Aignan)))


# Utilisation de la fonction

summary.commune(df_Nantes)

summary.commune(df_Aignan)

}
