% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.departement.R
\name{summary.departement}
\alias{summary.departement}
\title{Résumer les informations d'un département}
\usage{
\method{summary}{departement}(df, ...)
}
\arguments{
\item{df}{Un DataFrame contenant des informations sur les élus,
incluant les colonnes \verb{Libellé du département}, \verb{Code de la commune}, et les informations nécessaires
pour calculer la distribution des âges, trouver l’élu(e) le/la plus âgé(e) et le/la plus jeune,
ainsi que la moyenne d’âge par commune.}

\item{...}{Arguments supplémentaires, non utilisés dans cette méthode.}
}
\value{
Aucun retour explicite. La fonction affiche dans la console :
\itemize{
\item Le nom du département,
\item Le nombre de communes dans le département,
\item Le nombre d’élus dans le département,
\item La distribution des âges des élus sous forme de tibble,
\item Le nom, le prénom, la date de naissance et l'âge de l’élu(e) le/la plus âgé(e),
ainsi que celui/ceux de l’élu(e) le/la plus jeune, sous forme de tibble,
\item Les noms des communes avec la moyenne d’âge la plus faible et la plus élevée,
ainsi que la distribution des âges des élus pour ces communes.
}
}
\description{
Cette fonction affiche un résumé des informations d'un département,
notamment le nom du département, le nombre de communes, le nombre d'élus,
la distribution des âges des élus, ainsi que le nom et l'âge de l'élu(e) le/la plus âgé(e)
et de l'élu(e) le/la plus jeune. Elle affiche également les communes ayant la moyenne d’âge
la plus faible et la plus élevée, avec la distribution des âges des élus pour ces communes.
}
\examples{
# Exemples d'utilisation


# Données (départements)

df_Loire_Atlantique <- df_gers_loire_atlantique |>
  filter(`Libellé du département` == "Loire-Atlantique")

df_Gers <- df_gers_loire_atlantique |>
  filter(`Libellé du département` == "Gers")


# DataFrames appartenant à la classe departement

df_Loire_Atlantique <- structure(df_Loire_Atlantique, class = c("departement", class(df_Loire_Atlantique)))

df_Gers <- structure(df_Gers, class = c("df_Gers", class(df_Gers)))


# Utilisation de la fonction

summary.departement(df_Loire_Atlantique)

summary.departement(df_Gers)

}
