% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver_l_elu_le_plus_age.R
\name{trouver_l_elu_le_plus_age}
\alias{trouver_l_elu_le_plus_age}
\title{Trouver l'élu le plus âgé}
\usage{
trouver_l_elu_le_plus_age(df)
}
\arguments{
\item{df}{Un data frame contenant une colonne \verb{Date de naissance} au format \code{"jour/mois/année"},
ainsi que les colonnes \verb{Nom de l'élu} et \verb{Prénom de l'élu}.}
}
\value{
Un data frame avec les colonnes :
\itemize{
\item \verb{Nom de l'élu} : Nom de l'élu le plus âgé.
\item \verb{Prénom de l'élu} : Prénom de l'élu le plus âgé.
\item \verb{Date de naissance} : Date de naissance de l'élu le plus âgé.
\item \code{âge} : Âge en années de l'élu le plus âgé.
}
}
\description{
Cette fonction identifie l'élu le plus âgé dans un data frame,
en calculant l'âge à partir de la date de naissance.
}
\examples{
# Exemples d'utilisation


# Données (villes ou départements)

df_Nantes <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Nantes")

df_Aignan <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Aignan")

df_Loire_Atlantique <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Loire-Atlantique")

df_Gers <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Gers")


# Utilisation de la fonction

trouver_l'élu_le_plus_âgé(df_Nantes)

trouver_l'élu_le_plus_âgé(df_Aignan)

trouver_l'élu_le_plus_âgé(df_Loire_Atlantique)

trouver_l'élu_le_plus_âgé(df_Gers)
}
\keyword{internal}
