% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valider_schema.R
\name{valider_schema}
\alias{valider_schema}
\title{Valider le schema des donnees d'un data frame}
\usage{
valider_schema(df)
}
\arguments{
\item{df}{Un data frame contenant des informations sur les elus, qui doit avoir les colonnes suivantes :
\verb{Code du département}, \verb{Libellé du département}, \verb{Code de la collectivité à statut particulier},
\verb{Libellé de la collectivité à statut particulier}, \verb{Code de la commune}, \verb{Libellé de la commune},
\verb{Nom de l'élu}, \verb{Prénom de l'élu}, \verb{Code sexe}, \verb{Date de naissance}, \verb{Code de la catégorie socio-professionnelle},
\verb{Libellé de la catégorie socio-professionnelle}, \verb{Date de début du mandat}, \verb{Libellé de la fonction},
\verb{Date de début de la fonction}, et \verb{Code nationalité}.}
}
\value{
Aucun retour explicite. Si les colonnes du data frame ne correspondent pas au schema, une erreur sera levee.
}
\description{
Cette fonction verifie que les colonnes du data frame sont conformes à un schema predefini. Le schema attendu inclut des colonnes specifiques
liees aux informations sur les elus, telles que le code et le libelle du département, le code sexe, les dates de naissance, et d'autres informations
sur les elus et leurs fonctions.
}
\examples{
# Exemples d'utilisation


# Donnees (villes ou departements)

data("df_Gers_Loire_Atlantique")

df_Nantes <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Nantes")

df_Aignan <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Aignan")

df_Loire_Atlantique <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Loire-Atlantique")

df_Gers <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Gers")


# Utilisation de la fonction

valider_schema(df_Nantes)

valider_schema(df_Aignan)

valider_schema(df_Loire_Atlantique)

valider_schema(df_Gers)

valider_schema(df_Gers_Loire_Atlantique)
}
