% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valider_schema.R
\name{valider_schema}
\alias{valider_schema}
\title{Valider le schéma des données d'un data frame}
\usage{
valider_schema(df)
}
\arguments{
\item{df}{Un data frame contenant des informations sur les élus, qui doit avoir les colonnes suivantes :
\verb{Code du département}, \verb{Libellé du département}, \verb{Code de la collectivité à statut particulier},
\verb{Libellé de la collectivité à statut particulier}, \verb{Code de la commune}, \verb{Libellé de la commune},
\verb{Nom de l'élu}, \verb{Prénom de l'élu}, \verb{Code sexe}, \verb{Date de naissance}, \verb{Code de la catégorie socio-professionnelle},
\verb{Libellé de la catégorie socio-professionnelle}, \verb{Date de début du mandat}, \verb{Libellé de la fonction},
\verb{Date de début de la fonction}, et \verb{Code nationalité}.}
}
\value{
Aucun retour explicite. Si les colonnes du data frame ne correspondent pas au schéma, une erreur sera levée.
}
\description{
Cette fonction vérifie que les colonnes du data frame sont conformes à un schéma prédéfini. Le schéma attendu inclut des colonnes spécifiques
liées aux informations sur les élus, telles que le code et le libellé du département, le code sexe, les dates de naissance, et d'autres informations
sur les élus et leurs fonctions.
}
\examples{
# Exemples d'utilisation


# Données (villes ou départements)

data("df_Gers_Loire_Atlantique")

df_Nantes <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Nantes")

df_Aignan <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Aignan")

df_Loire_Atlantique <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Loire-Atlantique")

df_Gers <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Gers")


# Utilisation de la fonction

valider_schema(df_Nantes)

valider_schema(df_Aignan)

valider_schema(df_Loire_Atlantique)

valider_schema(df_Gers)

valider_schema(df_Gers_Loire_Atlantique)
}
